+++ src/ui/view/form/mod.rs
@@ -1,7 +1,7 @@
 // devela/src/ui/view/form/mod.rs
 //
 //! Cell, graphic, and document forms of UI projection.
-// #![doc = crate::_DOC_UI_VIEW_FORM!()] // public
+// #![doc = crate::_DOC_UI_VIEW_FORM!()] // private
 // #![doc = crate::_doc!(modules: crate::ui::view; form)]
 // #![doc = crate::_doc!(flat:"ui")]
 // #![doc = crate::_doc!(hr)]
@@ -9,6 +9,7 @@
 
 // mod cell;
 // mod document;
+// mod form; // UiViewForm
 // mod graphic;
 
 crate::structural_mods! { // _mods
@@ -16,6 +17,7 @@ crate::structural_mods! { // _mods
         // pub use super::{
         //     cell::_all::*,
         //     document::_all::*,
+        //     form::*,
         //     graphic::_all::*,
         // };
     }